+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # tsconfig
 
+## 0.11.11-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.11.10
 
 ### Patch Changes